chore: refactoring tests structure
