+++ 0000/day00/input.txt
