+++ 0000/day00/problem1.py
@@ -1,0 +1,2 @@
+def problem1(input):
+    return int(input[0]) * 1
