+++ 0000/day00/problem2.py
@@ -1,0 +1,2 @@
+def problem2(input):
+    return int(input[0]) * 2
