+++ 0000/day00/sample-input.txt
@@ -1,0 +1,1 @@
+1
