--- day00/problem1.py
@@ -1,2 +1,0 @@
-def problem1(input):
-    return input
