--- day00/problem2.py
@@ -1,2 +1,0 @@
-def problem2(input):
-    return input * 2
