--- day00/sample-input.txt
@@ -1,1 +1,0 @@
-42
